requirements.txt
torch
torchvision
numpy
Pillow
python-dotenv
matplotlib
ipywidgets
gradio
